In strerams, be mindful about having to await computations.
Make an empty Tree carry the Done signal (that still has to be awaited).
Make a completed Pollable carry the Done signal (that still has to be awaited).
Require that a closed Subscriber carries the Need signal (i.e. a closed Subscriber
still offers to take over the responsibility to await a Done signal).