Merge ClosedDSL and InvertDSL into Puro.
Rationale: It's unlikely that Puro (previously CoreDSL) would be used as a basis
for a different DSL than Libretto's concurrency DSL (as it would be too big of a basis).
With regards to the concurrency DSL, let's fully embrace closedness and inversions
as integral part of the DSL. It is unlikely that we would find a compelling case
to exclude them (e.g. for the sake of simpler implementation).